# [ci] speed up conda setup for macOS and Linux jobs